requirements.txt
attrs==21.2.0
blessed==1.17.6
boto3==1.18.29
botocore==1.21.29
certifi==2021.5.30
chalice==1.24.2
charset-normalizer==2.0.4
click==8.0.1
greenlet==1.1.2
idna==3.2
inquirer==2.7.0
jmespath==0.10.0
mypy-extensions==0.4.3
psycopg2-binary==2.9.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-editor==1.0.4
PyYAML==5.4.1
readchar==2.0.1
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
SQLAlchemy==1.4.26
urllib3==1.26.6
wcwidth==0.2.5
